requirements.txt
openai-agents
python-dotenv
pypdf
python-docx
rich
requests
